Add delete modal

+++ Client/src/Components/details/Details.jsx
@@ -1,18 +1,24 @@
 import { useState, useEffect, useContext } from "react";
 import { useNavigate, useParams, Link } from "react-router-dom";
+
+import { AuthContext } from "../../contexts/AuthContext.jsx";
+
 import styles from "./Details.module.css";
 import { deleteBook, getOneBook } from "../../services/bookService.js";
 import { buyBook, getBookSales, hasAlreadyBought } from "../../services/buyService.js";
-import { AuthContext } from "../../contexts/AuthContext.jsx";
+import DeleteModal from "./DeleteModal.jsx";
 
 const Details = () => {
     const { bookId } = useParams();
+    const navigate = useNavigate();
+
     const [book, setBook] = useState({});
     const [sales, setSales] = useState('');
     const [isOwner, setIsOwner] = useState(false);
-    const navigate = useNavigate();
+    const [hasBought, setHasBought] = useState('');
+    const [showModal, setShowModal] = useState(false);
     const { isAuthenticated, userId } = useContext(AuthContext);
-    const [ hasBought, setHasBought ] = useState('');
+
 
     useEffect(() => {
         getOneBook(bookId)
@@ -35,24 +41,39 @@ const Details = () => {
     }, [bookId])
 
     const onDelete = (bookId) => {
-        deleteBook(bookId)
-            .then(response => {
-                console.log(response);
-                if (response == 200) {
-                    navigate('/catalog')
-                } else {
-                    navigate('/catalog')
-                    // throw new Error('Unsuccessful delete');
-                }
-            })
-            .catch(error => {
-                console.log(error);
-            });
+        if (isOwner) {
+            deleteBook(bookId)
+                .then(response => {
+                    if (response == 200) {
+                        navigate('/catalog')
+                    } else {
+                        navigate('/catalog')
+                        // throw new Error('Unsuccessful delete');
+                    }
+                })
+                .catch(error => {
+                    console.log(error);
+                });
+        } else {
+            navigate('/catalog')
+        }
     }
 
+    const handleDeleteClick = (bookId) => {
+        setShowModal(true);
+    };
 
+    const handleCloseModal = () => {
+        setShowModal(false);
+    };
 
     const onBuy = (bookId) => {
+        if (isOwner) {
+            navigate('/catalog')
+        }
+        if (hasAlreadyBought > 0) {
+            navigate('/catalog')
+        }
         buyBook(bookId)
         setHasBought(1)
     }
@@ -73,15 +94,25 @@ const Details = () => {
                         <p>Sales: {sales}</p>
                     )}
                     <div className={styles.buttonContainer}>
-                        {!isOwner && hasBought<1 && (
+                        {!isOwner && hasBought < 1 && (
                             <button onClick={() => onBuy(bookId)}>Buy</button>
                         )}
-                        {hasBought>0 && (
+                        {hasBought > 0 && (
                             <p className={styles.sb}>Successfuly buying</p>
                         )}
                         {isOwner && (
                             <>
-                                <button onClick={() => onDelete(bookId)}>Delete</button>
+                                <button onClick={() => handleDeleteClick(bookId)}>Delete</button>
+
+                                {showModal && (
+                                    <DeleteModal
+                                        onDelete={onDelete}
+                                        bookId={bookId}
+                                        onClose={handleCloseModal}
+                                    />
+                                )}
+
+
                                 <Link to={`/edit/${bookId}`}> <button>Edit</button></Link>
                             </>
                         )}
